ft: improving project file management using barrels and re exports
