+++ src/store/setup.js
@@ -1,8 +1,8 @@
 import { configureStore } from '@reduxjs/toolkit'
 
-import recipesReducer from './recipeSlice';
-import userReducer from './userSlice';
-import notificationReducer from './notificationSlice';
+import recipesReducer from './recipe';
+import userReducer from './user';
+import notificationReducer from './notification';
 
 const store = configureStore({
   reducer: {
